+++ packages/api/src/ontology/PrimaryKeyTypes.ts
@@ -18,7 +18,6 @@ export type PrimaryKeyTypes =
   | "string"
   | "datetime"
   | "double"
-  | "boolean"
   | "integer"
   | "timestamp"
   | "short"