+++ packages/e2e.generated.catchall/src/generatedNoCheck/ontology/actions.ts
@@ -6,6 +6,7 @@ export { createMediaViaFunction } from './actions/createMediaViaFunction.js';
 export { createOsdkTestObject } from './actions/createOsdkTestObject.js';
 export { createStructPerson } from './actions/createStructPerson.js';
 export { createStructPersonOpiTeam } from './actions/createStructPersonOpiTeam.js';
+export { createTestGeoAction } from './actions/createTestGeoAction.js';
 export { createTodo } from './actions/createTodo.js';
 export { deleteFooInterface } from './actions/deleteFooInterface.js';
 export { deleteOsdkTestObject } from './actions/deleteOsdkTestObject.js';