+++ packages/widget.vite-plugin.unstable/src/dev-plugin/FoundryWidgetDevPlugin.ts
@@ -33,7 +33,12 @@ import {
 import { extractWidgetConfig } from "../common/extractWidgetConfig.js";
 import { getInputHtmlEntrypoints } from "../common/getInputHtmlEntrypoints.js";
 import { standardizeFileExtension } from "../common/standardizeFileExtension.js";
+import {
+  getCodeWorkspacesBaseHref,
+  isCodeWorkspacesMode,
+} from "./codeWorkspacesMode.js";
 import { extractInjectedScripts } from "./extractInjectedScripts.js";
+import { getFoundryToken } from "./getFoundryToken.js";
 import { getWidgetIdOverrideMap } from "./getWidgetIdOverrideMap.js";
 import { publishDevModeSettings } from "./publishDevModeSettings.js";
 
@@ -62,6 +67,17 @@ export function FoundryWidgetDevPlugin(): Plugin {
      */
     buildStart(options) {
       htmlEntrypoints = getInputHtmlEntrypoints(options);
+    },
+
+    /**
+     * Check for the required token environment variable in dev mode.
+     */
+    config(resolvedConfig, { command }) {
+      // Only check for the token environment variable when running in dev mode.
+      // When command is "serve" and not in test mode (VITEST).
+      if (command === "serve" && process.env.VITEST == null) {
+        getFoundryToken(resolvedConfig.mode);
+      }
     },
 
     /**
@@ -143,7 +159,12 @@ export function FoundryWidgetDevPlugin(): Plugin {
             configFiles,
             getLocalhostUrl(server),
           );
-          await publishDevModeSettings(server, widgetIdToOverrides, res);
+          await publishDevModeSettings(
+            server,
+            widgetIdToOverrides,
+            getBaseHref(server),
+            res,
+          );
         },
       );
 
@@ -220,6 +241,14 @@ function getLocalhostUrl(server: ViteDevServer): string {
   return `${
     server.config.server.https ? "https" : "http"
   }://localhost:${server.config.server.port}`;
+}
+
+function getBaseHref(
+  server: ViteDevServer,
+): string {
+  return isCodeWorkspacesMode(server.config.mode)
+    ? getCodeWorkspacesBaseHref()
+    : `${getLocalhostUrl(server)}${server.config.base}`;
 }
 
 /**