+++ packages/create-app.template-packager/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template-packager
 
+## 2.2.0-beta.16
+
+## 2.2.0-beta.15
+
 ## 2.2.0-beta.14
 
 ## 2.2.0-beta.13