+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @osdk/create-app
 
+## 2.2.0-beta.16
+
+## 2.2.0-beta.15
+
+### Minor Changes
+
+- 1444362: Read environment variables from index.html meta tags
+
 ## 2.2.0-beta.14
 
 ## 2.2.0-beta.13