+++ packages/generator-converters/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/generator-converters
 
+## 2.2.0-beta.16
+
+### Patch Changes
+
+- @osdk/api@2.2.0-beta.16
+
+## 2.2.0-beta.15
+
+### Patch Changes
+
+- Updated dependencies [ed69666]
+  - @osdk/api@2.2.0-beta.15
+
 ## 2.2.0-beta.14
 
 ### Patch Changes