+++ packages/maker/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @osdk/maker
 
+## 0.10.0-beta.16
+
+### Minor Changes
+
+- 4a02a1f: Make string more configurable
+- c0aa106: Add visibility field to shared property type in maker
+- 395ae07: Actions as code
+- c7f7a99: Export defineOntology method
+
+### Patch Changes
+
+- @osdk/api@2.2.0-beta.16
+
+## 0.10.0-beta.15
+
+### Minor Changes
+
+- 24c4b12: Add geotime and mediaset datasources to OAC
+- cc380b5: Use the latest Platform APIs
+
+### Patch Changes
+
+- Updated dependencies [ed69666]
+  - @osdk/api@2.2.0-beta.15
+
 ## 0.10.0-beta.14
 
 ### Minor Changes